requirements.txt
python-dotenv
google-generativeai
groq
rich
openai>=1.0.0
google-genai
google-cloud-aiplatform>=1.40.0
google-auth>=2.25.0
